ENH: Display fractional FPS

See merge request iMSTK/iMSTK!311

+++ Source/Scene/imstkScene.h
@@ -169,8 +169,8 @@ public:
     ///
     /// \brief Set/Get the FPS
     ///
-    void setFPS(const size_t fps) { m_fps = fps; }
-    size_t getFPS() { return m_fps; }
+    void setFPS(const double fps) { m_fps = fps; }
+    double getFPS() { return m_fps; }
 
 protected:
 
@@ -184,7 +184,7 @@ protected:
     std::vector<std::shared_ptr<SolverBase>> m_solvers;     ///> List of non-linear solvers
     std::vector<std::shared_ptr<SceneObjectControllerBase>> m_objectControllers; ///> List of controllers
 
-    size_t m_fps = 0;
+    double m_fps = 0.0;
 
     bool m_isInitialized = false;
 };
